requirements.txt
# Deep Learning Framework
torch==2.1.0
torchvision==0.16.0

# Interface Web
streamlit==1.29.0

# Visualização e Análise
matplotlib==3.8.2
seaborn==0.13.0
numpy==1.26.2
pandas==2.1.4

# Manipulação de Imagens
Pillow==10.1.0

# Utilidades
tqdm==4.66.1
scikit-learn==1.3.2

# Jupyter (para exploração)
jupyter==1.0.0
ipykernel==6.27.1
